+++ examples/src/examples/controls/OverviewMap.src.js
@@ -1,0 +1,28 @@
+import Map from 'ol/Map';
+import View from 'ol/View';
+import { defaults as defaultControls } from 'ol/control';
+import TileLayer from 'ol/layer/Tile';
+import OSM from 'ol/source/OSM';
+
+import OverviewMap from 'webgis4u/ol/control/OverviewMap';
+
+/**
+ * Initializes the map
+ */
+export function initialize(mapConfig) {
+  const map = new Map({
+    ...mapConfig,
+    controls: defaultControls().extend([
+      new OverviewMap(),
+    ]),
+    layers: [
+      new TileLayer({
+        source: new OSM(),
+      }),
+    ],
+    view: new View({
+      center: [0, 0],
+      zoom: 2,
+    }),
+  });
+}