+++ examples/src/examples/controls/index.js
@@ -1,1 +1,2 @@
 export { default as PanBar } from './PanBar';
+export { default as OverviewMap } from './OverviewMap';