+++ src/webgis4u/util.js
@@ -6,5 +6,6 @@ export * from './util/debounce';
 export * from './util/dom';
 export * from './util/formatNumber';
 export * from './util/formatText';
+export * from './util/ol';
 export * from './util/pan';
 export * from './util/string';