+++ CHANGELOG.md
@@ -10,3 +10,7 @@ All notable changes to this project will be documented in this file.
   * [`LayerRadioButton`](./src/webgis4u/ol/control/LayerRadioButton.js)
   * [`LegendDisplayable`](./src/webgis4u/ol/control/LegendDisplayable.js)
   * [`LegendVisibility`](./src/webgis4u/ol/control/LegendVisibility.js)
+  * [`OverviewMap`](./src/webgis4u/ol/control/OverviewMap.js)
+* Added utils:
+  * [`ol`](./src/webgis4u/util/ol.js):
+    * [`getMapChildElementByClassName`](./src/webgis4u/util/ol/getMapChildElementByClassName.js):